requirements.txt
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
cached-property==1.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
colorama==0.4.5
decorator==5.1.1
executing==0.10.0
greenlet==1.1.3
idna==3.3
ipdb==0.13.9
ipython==8.4.0
isodate==0.6.1
jedi==0.18.1
lxml==4.9.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.30
psycopg2==2.9.3
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
Pygments==2.13.0
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.9.1
sentry-sdk==1.9.8
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
stack-data==0.4.0
toml==0.10.2
traitlets==5.3.0
urllib3==1.26.12
wcwidth==0.2.5
yapf==0.32.0
zeep==4.1.0
